Moved all functions in Main.tsx to own util files
